requirements.txt
sentence-transformers
langchain
torch
pypdf
python-dotenv
pinecone-client
langchain-pinecone
langchain-community
langchain-google-genai
langchain-experimental
langchain-huggingface
langgraph
duckduckgo-search
streamlit
# -e .